Fix for GEOS-852, have JAI memory be a percentage of total memmory
git-svn-id: http://svn.codehaus.org/geoserver/branches/1.5.x@6016 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 2d66d81e8da6ed1e84da05126534b40adac6b87f [formerly 3b8414f347c926f0dcb191de93a18a17cf9afd7b]
Former-commit-id: 349399a7ec0c33d3387d461e4d97f65b78a0b16c

+++ geoserver/main/src/main/java/org/vfny/geoserver/global/xml/XMLConfigReader.java
@@ -516,11 +516,11 @@ public class XMLConfigReader {
                 }
             }
 
-            long jaiMemoryCapacity = 0;
+            double jaiMemoryCapacity = 0;
             elem = ReaderUtils.getChildElement(globalElem, "JaiMemoryCapacity", false);
 
             if (elem != null) {
-                jaiMemoryCapacity = (long) ReaderUtils.getDoubleAttribute(elem, "value", false);
+                jaiMemoryCapacity = ReaderUtils.getDoubleAttribute(elem, "value", false);
             }
 
             double jaiMemoryThreshold = 0.0;
